% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aloft_data_coverage.R
\name{aloft_data_coverage}
\alias{aloft_data_coverage}
\title{Get coverage table for Aloft bucket}
\usage{
aloft_data_coverage(use_cache = TRUE)
}
\arguments{
\item{use_cache}{Logical indicating whether to use the cache. Default is
\code{TRUE}. If \code{FALSE} the cache is ignored and the file is fetched from the
Aloft bucket. This can be useful if you want to force a refresh of the
cache.}
}
\value{
A data.frame of the coverage file on the Aloft bucket.
}
\description{
Retrieves an overview of the number of hdf5 files per directory in the
\href{https://aloftdata.eu/browse/}{Aloft bucket}, based on the \code{coverage.csv}
file in that bucket. By default this file is cached for 6 hours.
}
\details{
The coverage file is fetched from \url{https://aloftdata.s3-eu-west-1.amazonaws.com/coverage.csv}. This can be changed by
setting \code{options(getRad.aloft_data_url)} to any desired url.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
aloft_data_coverage()
\dontshow{\}) # examplesIf}
}
