% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pvol.R
\name{get_pvol}
\alias{get_pvol}
\title{Get polar volume (PVOL) data from supported sources}
\usage{
get_pvol(radar = NULL, datetime = NULL, ...)
}
\arguments{
\item{radar}{Name of the radar (odim code) as a character string (e.g.
\code{"nlhrw"} or \code{"fikor"}).}

\item{datetime}{Either:
\itemize{
\item A single \code{\link[base:DateTimeClasses]{POSIXct}}, for which the most
representative data file is downloaded. In most cases this will be the time
before.
\item A \code{\link[lubridate:interval]{lubridate::interval()}} or two \code{\link[base:DateTimeClasses]{POSIXct}},
between which all data files are downloaded.
}}

\item{...}{Additional arguments used to select reading functions or passed on
to reading functions. For example:
\itemize{
\item \code{use_opera_ord = TRUE} if you want to force using the
Opera open radar data over specific reading functions. Generally speaking
this data is less suitable for biological analysis. Note this data is only
available for 24 hours.
\item \code{param = c("DBZH", "VRADH")} to the \code{\link[bioRad:read_pvolfile]{bioRad::read_pvolfile()}}.
}}
}
\value{
Either a polar volume or a list of polar volumes. See
\code{\link[bioRad:summary.pvol]{bioRad::summary.pvol()}} for details.
}
\description{
Gets polar volume data from supported sources and returns it as a (list of)
\link[bioRad:summary.pvol]{polar volume objects}. The source is automatically
detected based on the provided \code{radar}.
}
\details{
For more details on supported sources, see \code{vignette("supported_sources")}. Within
supported countries there might also be temporal restrictions on the radars that
are operational. For example, radars with the \code{status} \code{0} in \code{get_weather_radars("opera")}
are currently not operational.

For European countries where a dedicated archive exist this data is used. Alternatively
data is downloaded from the Opera open radar data archive. This data is frequently
filtered. Use the argument \code{use_opera_ord = TRUE} if you want to force using the
Opera open radar data.

Not all radars in the nexrad archive can be read successfully. Radars associated
with the Terminal Doppler Weather Radar (TDWR) program can not be read. These can
be identified using the \code{stntype} column in \code{get_weather_radars("nexrad")}.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
# Get PVOL data for a single radar and datetime
get_pvol("deess", as.POSIXct(Sys.Date()))

# Get similar data from the opera open radar data
get_pvol("deess", as.POSIXct(Sys.Date()), use_opera_ord=TRUE)

# Get PVOL data for multiple radars and a single datetime
get_pvol(
  c("deess", "dehnr", "fianj", "czska", "KABR"),
  as.POSIXct(Sys.Date())
)

get_pvol("fianj",
  as.POSIXct("2012-05-11 23:00", tz = "UTC"),
  param = "DBZH"
)
\dontshow{\}) # examplesIf}
}
