% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vpts_coverage.R
\name{get_vpts_coverage}
\alias{get_vpts_coverage}
\title{Get coverage for vpts from various sources}
\usage{
get_vpts_coverage(source = c("baltrad", "uva", "ecog-04003", "rmi"))
}
\value{
A \code{data.frame} or \code{tibble}
}
\description{
Get coverage for vpts from various sources
}
\examples{
get_vpts_coverage()
}
