% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vpts_rmi.R
\name{get_vpts_rmi}
\alias{get_vpts_rmi}
\title{Get VPTS data from RMI}
\usage{
get_vpts_rmi(radar_odim_code, rounded_interval)
}
\arguments{
\item{radar_odim_code}{Radar ODIM code.}

\item{rounded_interval}{Interval to fetch data for, rounded to nearest day.}
}
\value{
A tibble with VPTS data.
}
\description{
Get VPTS data from \href{https://opendata.meteo.be/geonetwork/srv/eng/catalog.search#/metadata/RMI_DATASET_CROW}{RMI_DATASET_CROW}.
}
\keyword{internal}
