% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRad-package.R
\docType{package}
\name{getRad-package}
\alias{getRad}
\alias{getRad-package}
\title{getRad: Download Radar Data for Biological Research}
\description{
Load polar volume and vertical profile data for aeroecological research directly into R. With 'getRad' you can access data from several sources in Europe and the US and standardize it to facilitate further exploration in tools such as 'bioRad'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/aloftdata/getRad}
  \item \url{https://aloftdata.github.io/getRad/}
  \item Report bugs at \url{https://github.com/aloftdata/getRad/issues}
}

}
\author{
\strong{Maintainer}: Bart Kranstauber \email{b.kranstauber@uva.nl} (\href{https://orcid.org/0000-0001-8303-780X}{ORCID}) (University of Amsterdam)

Authors:
\itemize{
  \item Pieter Huybrechts \email{pieter.huybrechts@inbo.be} (\href{https://orcid.org/0000-0002-6658-6062}{ORCID}) (Research Institute for Nature and Forest (INBO))
  \item Peter Desmet \email{peter.desmet@inbo.be} (\href{https://orcid.org/0000-0002-8442-8025}{ORCID}) (Research Institute for Nature and Forest (INBO))
}

Other contributors:
\itemize{
  \item Cecilia Nilsson \email{cecilia.nilsson@biol.lu.se} (\href{https://orcid.org/0000-0001-8957-4411}{ORCID}) (Lund University) [contributor]
  \item Alexander Tedeschi \email{at744@cornell.edu} (\href{https://orcid.org/0000-0003-0772-6931}{ORCID}) (Cornell Lab of Ornithology) [contributor]
  \item Hidde Leijnse (\href{https://orcid.org/0000-0001-7835-4480}{ORCID}) (Royal Netherlands Meteorological Institute) [contributor]
  \item Bart Hoekstra (\href{https://orcid.org/0000-0002-7085-3805}{ORCID}) (University of Amsterdam) [contributor]
  \item University of Amsterdam (04dkp9463) [copyright holder]
  \item Biodiversa+ (https://hirad.science/) [funder]
}

}
\keyword{internal}
