% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_radars.R
\name{weather_radars}
\alias{weather_radars}
\title{Get weather radar metadata}
\usage{
weather_radars(source = c("opera"), use_cache = TRUE, ...)
}
\arguments{
\item{source}{A character vector for the one or more sources that should be downloaded,
values can be \code{"opera"},\code{"nexrad"}, and \code{"all"}}

\item{use_cache}{Logical indicating whether to use the cache. Default is
\code{TRUE}. If \code{FALSE} the cache is ignored and the file is fetched anew.
This can also be useful if you want to force a refresh of the cache.}

\item{...}{Arguments to helper functions specific to sources, currently not used}
}
\value{
A tibble with weather radar metadata. In all cases the column \code{source} is
added to indicate the source of the data and \code{radar} to show the radar identifiers
used in other functions like \code{\link[=get_pvol]{get_pvol()}} and \code{\link[=get_vpts]{get_vpts()}}.
}
\description{
Gets weather radar metadata from the \href{https://www.eumetnet.eu/activities/observations-programme/current-activities/opera/}{OPERA}
and/or \href{https://www.ncei.noaa.gov/products/radar/next-generation-weather-radar}{NEXRAD} network.
}
\details{
The source files for this function are:
\itemize{
\item For \code{opera}: \href{http://eumetnet.eu/wp-content/themes/aeron-child/observations-programme/current-activities/opera/database/OPERA_Database/OPERA_RADARS_DB.json}{OPERA_RADARS_DB.json}
and \href{http://eumetnet.eu/wp-content/themes/aeron-child/observations-programme/current-activities/opera/database/OPERA_Database/OPERA_RADARS_ARH_DB.json}{OPERA_RADARS_ARH_DB.json}.
\item For \code{nexrad} \href{https://www.ncei.noaa.gov/access/homr/file/nexrad-stations.txt}{nexrad-stations.txt}.
}
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
weather_radars()
\dontshow{\}) # examplesIf}
}
