% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aloft_data_coverage.R
\name{aloft_data_coverage}
\alias{aloft_data_coverage}
\title{Fetch the coverage table from the aloft data repository}
\usage{
aloft_data_coverage()
}
\value{
A data.frame of the coverage file on the aloft data repository
}
\description{
The coverage table provides an overview of what files are available on the
aloft data repository. Specifically it lists the directories that are
available as well as the number of files in every directory.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
aloft_data_coverage()
\dontshow{\}) # examplesIf}
}
