% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pvol.R
\name{get_pvol}
\alias{get_pvol}
\title{Retrieve polar volumes from \href{https://aloftdata.github.io/getRad/articles/supported_countries.html}{Supported countries}.}
\usage{
get_pvol(radar = NULL, datetime = NULL, ...)
}
\arguments{
\item{radar}{The name of the radar (odim string) as a character string (e.g.
\code{"nlhrw"} or \code{"fikor"}).}

\item{datetime}{The time as a \code{POSIXct} vector for the polar volume(s) to download or
an \code{\link[lubridate:interval]{interval}} in which all polar volumes need to be downloaded.}

\item{...}{Additional arguments passed on to the individual reading functions, for example \code{param="all"} to the \code{\link[bioRad:read_pvolfile]{bioRad::read_pvolfile()}} function.}
}
\value{
Either a polar volume or a list of polar volumes.
}
\description{
Retrieve polar volumes from \href{https://aloftdata.github.io/getRad/articles/supported_countries.html}{Supported countries}.
}
\details{
For more details on specific countries please see the map in the \href{https://aloftdata.github.io/getRad/articles/supported_countries.html}{vignette}.
}
\examples{
\dontrun{
get_pvol("deess", as.POSIXct(Sys.Date()))
get_pvol("czska", as.POSIXct(Sys.Date()))
get_pvol(
  c("deess", "dehnr", "fianj", "czska"),
  as.POSIXct(Sys.Date())
)
}
}
