% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pvol.R
\name{get_pvol}
\alias{get_pvol}
\title{Get polar volume (PVOL) data from supported countries}
\usage{
get_pvol(radar = NULL, datetime = NULL, ...)
}
\arguments{
\item{radar}{Name of the radar (odim code) as a character string (e.g.
\code{"nlhrw"} or \code{"fikor"}).}

\item{datetime}{Either:
\itemize{
\item A single \code{\link[base:DateTimeClasses]{POSIXct}}, for which the most representative data
file is downloaded. In most cases this will be the time before.
\item A \code{\link[lubridate:interval]{lubridate::interval()}} or two \code{\link[base:DateTimeClasses]{POSIXct}}, between which all data files with a reference time in the interval are downloaded.
}}

\item{...}{Additional arguments passed on to reading functions, for example
\code{param = "all"} to the \code{\link[bioRad:read_pvolfile]{bioRad::read_pvolfile()}}.}
}
\value{
Either a polar volume or a list of polar volumes. See
\code{\link[bioRad:summary.pvol]{bioRad::summary.pvol()}} for details.
}
\description{
Gets polar volume data for supported countries and returns it as a (list of)
\link[bioRad:summary.pvol]{polar volume objects}.
}
\details{
For more details on supported countries, see
\code{vignette("supported_countries")}.
}
\examples{
\dontrun{
get_pvol("deess", as.POSIXct(Sys.Date()))
get_pvol("czska", as.POSIXct(Sys.Date()))
get_pvol(
  c("deess", "dehnr", "fianj", "czska"),
  as.POSIXct(Sys.Date())
)
}
}
