% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pvol.R
\name{get_pvol}
\alias{get_pvol}
\title{Retrieve polar volumes}
\usage{
get_pvol(radar = NULL, time = NULL, ...)
}
\arguments{
\item{radar}{The name of the radar as a scalar character string.}

\item{time}{The time as a \code{POSIXct} for the polar volume to download}

\item{...}{Additional arguments passed on tho the individual reading functions, for example \code{param="all"} to the bioRad::read_pvolfile function.}
}
\value{
Either a polar volume or a list of polar volumes
}
\description{
Retrieve polar volumes
}
\details{
For more details on specific countries please see the vignettes
}
\examples{
\dontrun{
get_pvol("deess", as.POSIXct(Sys.Date()))
get_pvol("czska", as.POSIXct(Sys.Date()))
get_pvol(
  c("deess", "dehnr", "fianj", "czska"),
  as.POSIXct(Sys.Date())
)
}
}
