% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vpts.R
\name{get_vpts}
\alias{get_vpts}
\title{Retrieve vertical profile time series data from the Aloft bucket}
\usage{
get_vpts(
  radar,
  datetime,
  source = c("baltrad", "uva", "ecog-04003"),
  return_type = c("vpts", "tibble")
)
}
\arguments{
\item{radar}{The name of the radar (odim string) as a character string (e.g.
\code{"nlhrw"} or \code{"fikor"}).}

\item{datetime}{Either a single date or range as either a \code{character},
\code{\link[base:Dates]{Date}}, \code{\link[base:DateTimeClasses]{POSIXct}} or a
\link[lubridate:interval]{lubridate::interval}}

\item{source}{The source of the data. One of \code{"baltrad"}, \code{"uva"} or
\code{"ecog-04003"}. Only one source can be queried at a time. If no source is
provided \code{baltrad} is used.}

\item{return_type}{The type of object that should be returned. By default the
data is returned as a \code{\link[bioRad:summary.vpts]{vpts}} object. If set to
\code{"tibble"}, a \code{\link[dplyr:reexports]{dplyr::tibble()}} will be returned instead with an extra
column for the radar source.}
}
\value{
By default, a vpts object is returned. See \link[bioRad:summary.vpts]{bioRad::summary.vpts} for
more information. When multiple radars are selected, a list of vpts objects
will be returned instead. When \code{return_type = "tibble"}, a single
\code{\link[dplyr:reexports]{tibble}} is returned with an extra column for the radar
source.
}
\description{
Retrieve vertical profile time series data from the Aloft bucket
}
\details{
By default, data from the \href{https://aloftdata.eu/browse/}{Aloft bucket} is retrieved from
\url{https://aloftdata.s3-eu-west-1.amazonaws.com}, this can be changed by setting
\code{options(getRad.aloft_data_url)} to any desired url.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

  # Fetch vpts data for a single radar and date

  get_vpts(radar = "bejab", datetime = "2023-01-01", source = "baltrad")

  # Fetch vpts data for multiple radars and a single date

  get_vpts( radar = c("dehnr", "deflg"), datetime =
  lubridate::ymd("20171015"), source = "baltrad" )

  # Fetch vpts data for a single radar and a date range

  get_vpts( radar = "bejab", datetime = lubridate::interval(
  lubridate::ymd_hms("2023-01-01 00:00:00"), lubridate::ymd_hms("2023-01-02
  00:14:00") ), source = "baltrad" )

  get_vpts("bejab", lubridate::interval("20210101", "20210301"))

  # Fetch vpts data for a single radar and a date range from a specific #
  source

  get_vpts(radar = "bejab", datetime = "2016-09-29", source = "ecog-04003")

  # Return a tibble instead of a vpts object

get_vpts(
  radar = "chlem", date = "2023-03-10", source = "baltrad",
  return_type = "tibble"
)

  # get_vpts() can return data for a whole day

  get_vpts(radar = "sehem", datetime = "20170704")

  # If you provide an interval with specific time information, only for this
  interval is returned

  get_vpts(radar = "frtou", datetime = lubridate::interval("2021-01-17 22:15:00",
                                                           "2021-01-18 02:45:00"))
\dontshow{\}) # examplesIf}
}
