% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vpts.R
\name{get_vpts}
\alias{get_vpts}
\title{Retrieve vertical profile time series data from the Aloft data repository}
\usage{
get_vpts(
  radar,
  date,
  source = c("baltrad", "uva", "ecog-04003"),
  as_tibble = FALSE
)
}
\arguments{
\item{radar}{The name of the radar (odim string) as a character string (e.g. \code{"nlhrw"} or \code{"fikor"}).}

\item{date}{Either a single date or a \link[lubridate:interval]{lubridate::interval}}

\item{source}{The source of the data. One of baltrad, uva or ecog-04003. Only
one source can be queried at a time. This is a required argument.}

\item{as_tibble}{Logical. By default the data is returned as a
\link[bioRad:summary.vpts]{bioRad::summary.vpts} object. If set to TRUE, a \code{\link[dplyr:reexports]{dplyr::tibble()}} will be
returned instead with an extra column for the radar source.}
}
\value{
By default, a vpts object is returned. See \link[bioRad:summary.vpts]{bioRad::summary.vpts} for
more information. When multiple radars are selected, a list of vpts objects
will be returned instead. When \code{as_tibble = TRUE}, a single
\code{\link[dplyr:reexports]{dplyr::tibble()}} is returned with an extra column for the radar source.
}
\description{
Retrieve vertical profile time series data from the Aloft data repository
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

  # Fetch vpts data for a single radar and date
  get_vpts(radar = "bejab", date = "2023-01-01", source = "baltrad")

  # Fetch vpts data for multiple radars and a single date

  get_vpts(radar = c("dehnr", "deflg"), date = lubridate::ymd("20171015"),
  source = "baltrad")

  # Fetch vpts data for a single radar and a date range

  get_vpts(radar = "bejab", date = lubridate::interval(
  lubridate::ymd_hms("2023-01-01 00:00:00"), lubridate::ymd_hms("2023-01-02
  00:14:00") ), source = "baltrad" )

  get_vpts("bejab", lubridate::interval("20210101","20210301"), "bejab")

  # Fetch vpts data for a single radar and a date range from a specific
  source

  get_vpts(radar = "bejab", date = "2016-09-29", source = "ecog-04003")

  # Return a tibble instead of a vpts object

  get_vpts(radar = "chlem", date = "2023-03-10", source = "baltrad",
  as_tibble = TRUE)
\dontshow{\}) # examplesIf}
}
