% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vpts_rmi.R
\name{get_vpts_rmi}
\alias{get_vpts_rmi}
\title{Get VPTS RMI}
\usage{
get_vpts_rmi(radar_odim_code, rounded_interval)
}
\arguments{
\item{radar_odim_code}{The radar ODIM code.}

\item{rounded_interval}{The interval to fetch data for, rounded to nearest
day}
}
\value{
A list of vpts data.frames from RMI.
}
\description{
Get VPTS RMI
}
\examples{
get_vpts_rmi("bejab",
             lubridate::interval("20200119", "20200124"))
}
\keyword{internal}
