% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRad-package.R
\docType{package}
\name{getRad-package}
\alias{getRad}
\alias{getRad-package}
\title{getRad: Access and Standardize Radar Data}
\description{
This package facilitates downloading radar data for biological research. Currently the main functionality is downloading polar volume data from open repositories of European countries.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/enram/getRad}
  \item \url{https://enram.github.io/getRad/}
  \item Report bugs at \url{https://github.com/enram/getRad/issues}
}

}
\author{
\strong{Maintainer}: Bart Kranstauber \email{b.kranstauber@uva.nl} (\href{https://orcid.org/0000-0001-8303-780X}{ORCID}) (University of Amsterdam)

Authors:
\itemize{
  \item Peter Desmet \email{peter.desmet@inbo.be} (\href{https://orcid.org/0000-0002-8442-8025}{ORCID}) (Research Institute for Nature and Forest (INBO))
  \item Pieter Huybrechts \email{pieter.huybrechts@inbo.be} (\href{https://orcid.org/0000-0002-6658-6062}{ORCID}) (Research Institute for Nature and Forest (INBO))
}

}
\keyword{internal}
