% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secrets.R
\name{get_secret}
\alias{get_secret}
\alias{set_secret}
\title{Setting and retrieving secrets from the keyring}
\usage{
get_secret(name)

set_secret(name, secret = NULL)
}
\arguments{
\item{name}{Name of the secret to set or retrieve as a scalar character (e.g. \code{"nl_api_key"})}

\item{secret}{Optionally a character string the the secret alternatively the system will prompt the user

When working with a cluster it might be advantageous to use a specific keyring this can be done by setting the \code{keyring_backend} option in R.

The package uses the option \code{getRad.key_prefix} as a prefix to all keys stored. If you want to use multiple keys for the same api you can manipulate this option.}
}
\value{
\code{get_secret} returns a character string, \code{set_secret} \code{TRUE} if successful.
}
\description{
Setting and retrieving secrets from the keyring
}
