% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_radars.R
\name{weather_radars}
\alias{weather_radars}
\title{Get weather radar data}
\usage{
weather_radars(use_cache = TRUE)
}
\arguments{
\item{use_cache}{Logical indicating whether to use the cache. Default is
\code{TRUE}. If \code{FALSE} the cache is ignored and the file is fetched anew.
This can also be useful if you want to force a refresh of the cache.}
}
\value{
A tibble containing weather radar metadata
}
\description{
The \code{weather_radars()} function is used to retreive weather radar metadata
from the \href{https://www.eumetnet.eu/activities/observations-programme/current-activities/opera/}{OPERA network}.
The retrieved data is returned as a tibble.
}
\examples{
weather_radars()
}
