% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_radars.R
\name{weather_radars}
\alias{weather_radars}
\title{Get weather radar data}
\usage{
weather_radars()
}
\value{
A tibble containing weather radar metadata
}
\description{
The \code{weather_radars()} function is used to retreive weather radar metadata
from the \href{https://www.eumetnet.eu/activities/observations-programme/current-activities/opera/}{OPERA network}.
The retrieved data is returned as a tibble.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
weather_radars()
\dontshow{\}) # examplesIf}
}
